make z80pack more friendly for use on bare hardware
1. rename sim[0-7]*.c to make it clear what the files actually contain.
2. Add new file simcore.c which contains core CPU and bus handling functions.
3. Use run_cpu() and step_cpu() from simcore.c everywhere.
4. Synchronize mon() functions in altairsim, cromemcosim and imsaisim.
5. Rename sleep_ms() to do_sleep_ms() to avoid ARM SDK clash
6. Add CORE_LOG define to sim.h to tell if core simulator functions should
   use LOG() or not.
7. Add EXCLUDE_Z80 and EXCLUDE_I8080 options for use on bare hardware.

For a core 8080 simulator add EXCLUDE_Z80 to sim.h and use:
        simcore.c sim8080.c simglb.c

For a core Z80 simulator add EXCLUDE_I8080 to sim.h and use:
        simcore.c simz80.c simz80-cb.c simz80-dd.c simz80-ddcb.c
        simz80-ed.c simz80-fd.c simz80-fdcb.c simglb.c

That leaves SLEEP_MS, which is used by the CPU speed adjusting and HALT.
