+++ altairsim/srcsim/simctl.c
@@ -47,8 +47,8 @@
 #include "../../iodevices/unix_terminal.h"
 #include "log.h"
 
-extern void cpu_z80(void), cpu_8080(void);
 extern void reset_cpu(void), reset_io(void);
+extern void run_cpu(void), step_cpu(void);
 extern void report_cpu_error(void);
 
 static const char *TAG = "system";
@@ -62,11 +62,7 @@ static BYTE power_switch = 1;
 static int power;
 #endif
 
-#if defined(FRONTPANEL) || !defined(WANT_ICE)
-static void run_cpu(void);
-#endif
 #ifdef FRONTPANEL
-static void step_cpu(void);
 static void run_clicked(int, int), step_clicked(int, int);
 static void reset_clicked(int, int);
 static void examine_clicked(int, int), deposit_clicked(int, int);
@@ -129,8 +125,8 @@ void mon(void)
 	SLEEP_MS(999);
 	fflush(stdout);
 
-	/* initialise terminal */
 #ifndef WANT_ICE
+	/* initialise terminal */
 	set_unix_terminal();
 #endif
 	atexit(reset_unix_terminal);
@@ -138,6 +134,7 @@ void mon(void)
 #ifdef FRONTPANEL
 	/* operate machine from front panel */
 	while (cpu_error == NONE) {
+		/* update frontpanel LED's */
 		if (reset) {
 			cpu_bus = 0;
 			fp_led_address = 0xffff;
@@ -161,6 +158,7 @@ void mon(void)
 		fp_clock++;
 		fp_sampleData();
 
+		/* run CPU if not idling */
 		switch (cpu_switch) {
 		case 1:
 			if (!reset)
@@ -178,26 +176,35 @@ void mon(void)
 		fp_clock++;
 		fp_sampleData();
 
+		/* wait a bit, system is idling */
 		SLEEP_MS(10);
 	}
+
+	/* check for CPU emulation errors and report */
+	report_cpu_error();
 #else
 #ifdef WANT_ICE
 	extern void ice_cmd_loop(int);
 
 	ice_before_go = set_unix_terminal;
 	ice_after_go = reset_unix_terminal;
+	atexit(reset_unix_terminal);
+
 	ice_cmd_loop(0);
 #else
 	/* run the CPU */
 	run_cpu();
+
+	/* check for CPU emulation errors and report */
+	report_cpu_error();
 #endif
 #endif
 
 #ifndef WANT_ICE
 	/* reset terminal */
 	reset_unix_terminal();
-	putchar('\n');
 #endif
+	putchar('\n');
 
 #ifdef FRONTPANEL
 	/* all LED's off and update front panel */
@@ -217,46 +224,7 @@ void mon(void)
 #endif
 }
 
-#if defined(FRONTPANEL) || !defined(WANT_ICE)
-/*
- *	Run CPU
- */
-void run_cpu(void)
-{
-	cpu_state = CONTIN_RUN;
-	cpu_error = NONE;
-	switch (cpu) {
-	case Z80:
-		cpu_z80();
-		break;
-	case I8080:
-		cpu_8080();
-		break;
-	}
-	report_cpu_error();
-}
-#endif
-
 #ifdef FRONTPANEL
-/*
- *	Step CPU
- */
-void step_cpu(void)
-{
-	cpu_state = SINGLE_STEP;
-	cpu_error = NONE;
-	switch (cpu) {
-	case Z80:
-		cpu_z80();
-		break;
-	case I8080:
-		cpu_8080();
-		break;
-	}
-	cpu_state = STOPPED;
-	report_cpu_error();
-}
-
 /*
  *	Callback for RUN/STOP switch
  */
