+++ cpmsim/srcsim/Makefile
@@ -80,7 +80,8 @@ LDFLAGS = $(PLAT_LDFLAGS) $(MACHINE_LIBS) \
 	  $(PLAT_LIBS)
 
 # core system source files for the CPU simulation - only change if the core changes
-CORE_SRCS = sim0.c sim1.c sim1a.c sim2.c sim3.c sim4.c sim5.c sim6.c sim7.c simdis.c simfun.c simglb.c simice.c simint.c
+CORE_SRCS = sim8080.c simcore.c simdis.c simfun.c simglb.c simice.c simint.c simmain.c \
+	simz80.c simz80-cb.c simz80-dd.c simz80-ddcb.c simz80-ed.c simz80-fd.c simz80-fdcb.c
 SRCS = $(CORE_SRCS) $(MACHINE_SRCS) $(IO_SRCS)
 OBJS = $(SRCS:.c=.o)
