+++ cpmsim/srcsim/sim.h
@@ -12,6 +12,7 @@
 #define CPU_SPEED 0	/* default CPU speed 0=unlimited */
 #define Z80_UNDOC	/* compile undocumented Z80 instructions */
 #define WANT_FASTB	/* much faster but not accurate Z80 block instr. */
+#define CORE_LOG	/* use LOG() logging in core simulator */
 
 /*#define WANT_ICE*/	/* attach ICE to machine */
 /*#define WANT_TIM*/	/* don't count t-states */
