+++ imsaisim/srcsim/sim.h
@@ -33,6 +33,7 @@
 #define CPU_SPEED 2	/* default CPU speed */
 #define Z80_UNDOC	/* compile undocumented Z80 instructions */
 /*#define WANT_FASTB*/	/* much faster but not accurate Z80 block instr. */
+#define CORE_LOG	/* use LOG() logging in core simulator */
 
 /*#define WANT_ICE*/	/* attach ICE to headless machine */
 /*#define WANT_TIM*/	/* don't count t-states */
