+++ picosim/picosim.c
@@ -16,7 +16,8 @@
 #include "memory.h"
 
 extern void init_cpu(void);
-extern void cpu_z80(void);
+extern void run_cpu(void);
+extern void report_cpu_error(void);
 
 /*
  *	Compute difference between two timeval in microseconds
@@ -52,53 +53,12 @@ int main(void)
 
 	config();		/* read system configuration */
 	init_cpu();		/* initialise CPU */
-	init_memory();	/* initialise memory configuration */
+	init_memory();		/* initialise memory configuration */
 
-	cpu_state = CONTIN_RUN;
-	cpu_z80();		/* run the Z80 with whatever is in memory */
+	run_cpu();		/* run the CPU with whatever is in memory */
 
-	printf("\n");
-	switch (cpu_error) {
-	case NONE:
-		break;
-	case OPHALT:
-		printf("INT disabled and HALT Op-Code reached at %04x",
-		       PC - 1);
-		break;
-	case IOTRAPIN:
-		printf("I/O input Trap at %04x, port %02x", PC, io_port);
-		break;
-	case IOTRAPOUT:
-		printf("I/O output Trap at %04x, port %02x", PC, io_port);
-		break;
-	case IOHALT:
-		printf("System halted");
-		break;
-	case IOERROR:
-		printf("Fatal I/O Error at %04x", PC);
-		break;
-	case OPTRAP1:
-		printf("Op-code trap at %04x %02x", PC - 1,
-		       getmem(PC - 1));
-		break;
-	case OPTRAP2:
-		printf("Op-code trap at %04x %02x %02x",
-		       PC - 2, getmem(PC - 2),
-		       getmem(PC - 1));
-		break;
-	case OPTRAP4:
-		printf("Op-code trap at %04x %02x %02x %02x %02x",
-		       PC - 4, getmem(PC - 4), getmem(PC - 3),
-		       getmem(PC - 2), getmem(PC - 1));
-		break;
-	case INTERROR:
-		printf("Unsupported bus data during INT: %02x",
-		       int_data);
-		break;
-	default:
-		printf("Unknown error %d", cpu_error);
-		break;
-	}
-	printf("\nBye.\n\n");
+	putchar('\n');
+	report_cpu_error();	/* check for CPU emulation errors and report */
+	puts("Bye.\n");
 	stdio_flush();
 }
