+++ picosim/sim.h
@@ -9,5 +9,7 @@
 #define DEF_CPU Z80     /* default CPU (Z80 or I8080) */
 #define CPU_SPEED 0     /* default CPU speed 0=unlimited */
 #define Z80_UNDOC       /* compile undocumented Z80 instructions */
+/*#define CORE_LOG*/	/* don't use LOG() logging in core simulator */
+#define EXCLUDE_I8080	/* don't include 8080 emulation support */
 
 #include "simcore.h"
