+++ z80sim/srcsim/sim.h.fast
@@ -15,6 +15,7 @@
 #define CPU_SPEED 0	/* default CPU speed 0=unlimited */
 #define Z80_UNDOC	/* compile undocumented Z80 instructions */
 #define WANT_FASTB	/* much faster but not accurate Z80 block instr. */
+/*#define CORE_LOG*/	/* don't use LOG() logging in core simulator */
 
 /*#define WANT_TIM*/	/* count t-states */
 /*#define HISIZE 100*/	/* number of entries in history */
